+++ R/model_performance.mixed.R
@@ -3,7 +3,7 @@
 #' Compute indices of model performance for mixed models.
 #'
 #' @param model Object of class \code{merMod}, \code{glmmTMB}, \code{lme} or \code{MixMod}.
-#' @param metrics Can be \code{"all"} or a character vector of metrics to be computed (some of \code{c("AIC", "BIC", "R2", "ICC", "RMSE", "LOGLOSS")}).
+#' @param metrics Can be \code{"all"} or a character vector of metrics to be computed (some of \code{c("AIC", "BIC", "R2", "ICC", "RMSE", "LOGLOSS", "SCORE")}).
 #' @param ... Arguments passed to or from other methods.
 #'
 #' @return A data frame (with one row) and one column per "index" (see \code{metrics}).
@@ -11,6 +11,9 @@
 #' @details This method returns the \emph{adjusted ICC} only, as this is typically
 #'   of interest when judging the variance attributed to the random effects part
 #'   of the model (see also \code{\link{icc}}).
+#'   \cr \cr
+#'   Furthermore, see 'Details' in \code{\link{model_performance.lm}} for
+#'   more details on returned indices.
 #'
 #' @examples
 #' library(lme4)
@@ -22,7 +25,7 @@
 #' @export
 model_performance.merMod <- function(model, metrics = "all", ...) {
   if (all(metrics == "all")) {
-    metrics <- c("AIC", "BIC", "R2", "ICC", "RMSE", "LOGLOSS")
+    metrics <- c("AIC", "BIC", "R2", "ICC", "RMSE", "LOGLOSS", "SCORE")
   }
 
   mi <- insight::model_info(model)
@@ -47,6 +50,11 @@ model_performance.merMod <- function(model, metrics = "all", ...) {
   }
   if (("LOGLOSS" %in% metrics) && mi$is_binomial) {
     out$LOGLOSS <- performance_logloss(model)
+  }
+  if (("SCORE" %in% metrics) && (mi$is_binomial || mi$is_count)) {
+    .scoring_rules <- performance_score(model)
+    if (!is.na(.scoring_rules$logarithmic)) out$SCORE_LOG <- .scoring_rules$logarithmic
+    if (!is.na(.scoring_rules$spherical)) out$SCORE_SPHERICAL <- .scoring_rules$spherical
   }
 
   # TODO: What with sigma and deviance?